+++ CHANGELOG.md
@@ -38,6 +38,8 @@ Upcoming feature release, possibly around 1 August 2026.
  - Extra comma when printing `X_FLOAT` value to JSON.
  
  - Fixed JSON unicode processing.
+ 
+ - Fixed `PrintPrimitive()` return value for `X_CHAR(n)` types in `xjson.c`.
 
 ### Added